Revert "Merge branch 'STORM-2236' of https://github.com/MediaV/storm into STORM-2236"
This reverts commit e2f87c304297ce95f01708e0934c44ee385fa0bb, reversing
changes made to 00fed518d14b485e2a9e9cdd91ade3cc9d68dc8d.
